requirements.txt
Django==3.2.*
djangorestframework==3.14.0
requests
django-environ